analytics: Aggregate empty stream and message counts in RealmCount
The previous do_aggregate behavior for RealmCount differed from
InstallationCount in that if a realm had no users and no
messages/streams, we would not insert a row. This commit changes
that behavior so that we write a row with value 0 to RealmCount
if there are no messages or streams for a given realm. This behavior
is needed for analysis of messages or streams by realm that is
independent of user information.