analytics: Add counts for messages sent to PMs, streams, etc by user.
Similar in spirit to a count_recipient_by_user query, where we would join on
the Message, Recipient, and UserProfile table. In addition, we join on the
Stream table in order to distinguish private and public streams.We also
count PM and Huddle type messages as a single group, PMs.